- CodeManager removed

+++ builder/builder.cpp
@@ -2,15 +2,13 @@
 
 #include "log.h"
 #include "projectmanager.h"
-#include "codemanager.h"
 
 #include <QCoreApplication>
 
 Builder::Builder() {
-    connect(AssetManager::instance(), &AssetManager::importFinished, CodeManager::instance(), &CodeManager::rebuildProject);
+    connect(AssetManager::instance(), &AssetManager::importFinished, AssetManager::instance(), &AssetManager::rebuildProject);
 
-    connect(CodeManager::instance(), SIGNAL(buildSucess(QString)), this, SLOT(onCompileDone(QString)));
-    connect(CodeManager::instance(), &CodeManager::buildFailed, QCoreApplication::instance(), &QCoreApplication::quit);
+    connect(AssetManager::instance(), SIGNAL(buildFinished(int)), this, SLOT(onCompileFinished(int)));
 
     connect(this, &Builder::packDone, QCoreApplication::instance(), &QCoreApplication::quit);
     connect(this, &Builder::moveDone, this, &Builder::package);
@@ -99,18 +97,21 @@ bool copyRecursively(QString sourceFolder, QString destFolder) {
     return true;
 }
 
-void Builder::onCompileDone(const QString &path) {
-    QFileInfo info(path);
-    QFileInfo target(ProjectManager::instance()->targetPath() + "/" + info.fileName());
+void Builder::onCompileFinished(int code) {
+    if(code == 0) {
+        QString path = AssetManager::instance()->artifact();
+        QFileInfo info(path);
+        QFileInfo target(ProjectManager::instance()->targetPath() + "/" + info.fileName());
 
-    if((target.isDir() && QDir(target.absoluteFilePath()).removeRecursively()) || QFile::remove(target.absoluteFilePath())) {
-        Log(Log::INF) << "Previous build removed.";
-    }
+        if((target.isDir() && QDir(target.absoluteFilePath()).removeRecursively()) || QFile::remove(target.absoluteFilePath())) {
+            Log(Log::INF) << "Previous build removed.";
+        }
 
-    if((info.isDir() && copyRecursively(path, target.absoluteFilePath())) || QFile::copy(path, target.absoluteFilePath())) {
-        Log(Log::INF) << "New build copied to:" << qPrintable(target.absoluteFilePath());
-        emit moveDone(target.absoluteFilePath());
-        return;
+        if((info.isDir() && copyRecursively(path, target.absoluteFilePath())) || QFile::copy(path, target.absoluteFilePath())) {
+            Log(Log::INF) << "New build copied to:" << qPrintable(target.absoluteFilePath());
+            emit moveDone(target.absoluteFilePath());
+            return;
+        }
     }
-    QCoreApplication::exit(1);
+    QCoreApplication::exit(code);
 }
